add confirmation prompt to deletequeues commandlet

+++ Pipeline/TileServer/Commands/DeleteQueues.cs
@@ -8,6 +8,8 @@ namespace OPS.Pipeline.TileServer
     [Verb("deletequeues", HelpText = "Delete queues")]
     public class DeleteQueuesOptions
     {       
+        [Option(HelpText = "Disable confirmation prompt", Default = false)]
+        public bool Force { get; set; }
     }
 
     public class DeleteQueues : PipelineCore
@@ -25,7 +27,14 @@ namespace OPS.Pipeline.TileServer
         public int Run()
         {
             var cloud = new TileServerCloud(this);
-            logger.Info("WARNING deleting queues: " + cloud.MasterQueue.Name + ", " + cloud.WorkerQueue.Name);
+            string queues = cloud.MasterQueue.Name + ", " + cloud.WorkerQueue.Name;
+            if (!options.Force)
+            {
+                Console.WriteLine("delete queues " + queues + " (yes/no)?");
+                var response = Console.ReadLine();
+                if (response.ToLower() != "yes") return 1;
+            }
+            logger.Info("deleting queues: " + queues);
             cloud.DeleteQueues();
             return 0;
         }
